rename QuotientFamilial::VX
