+++ app/interactors/get_family_quotient.rb
@@ -14,6 +14,6 @@ class GetFamilyQuotient < BaseInteractor
   private
 
   def quotient_familial
-    @quotient_familial ||= ApiParticulier::QuotientFamilialV2.get(access_token: context.user.access_token, siret: context.siret)
+    @quotient_familial ||= ApiParticulier::QuotientFamilial::V2.get(access_token: context.user.access_token, siret: context.siret)
   end
 end
